+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+# v0.27.1: More cooling-related updates, based on additional info from Plugwise
+- Updates for Anna+Elga and Adam-OnOff systems
+- Loria/Thermastage fix
+
 # v0.27.0: Smile P1: support 3-phase measurements
 
 (# v0.26.0: not released)